move PriceData to /common
