+++ packages/xrpl/src/models/common/index.ts
@@ -155,3 +155,36 @@ export interface XChainBridge {
   IssuingChainDoor: string
   IssuingChainIssue: Currency
 }
+
+/**
+ * A PriceData object represents the price information for a token pair.
+ *
+ */
+export interface PriceData {
+  PriceData: {
+    /**
+     * The primary asset in a trading pair. Any valid identifier, such as a stock symbol, bond CUSIP, or currency code is allowed.
+     * For example, in the BTC/USD pair, BTC is the base asset; in 912810RR9/BTC, 912810RR9 is the base asset.
+     */
+    BaseAsset: string
+
+    /**
+     * The quote asset in a trading pair. The quote asset denotes the price of one unit of the base asset. For example, in the
+     * BTC/USD pair,BTC is the base asset; in 912810RR9/BTC, 912810RR9 is the base asset.
+     */
+    QuoteAsset: string
+
+    /**
+     * The asset price after applying the Scale precision level. It's not included if the last update transaction didn't include
+     * the BaseAsset/QuoteAsset pair.
+     */
+    AssetPrice?: number | string
+
+    /**
+     * The scaling factor to apply to an asset price. For example, if Scale is 6 and original price is 0.155, then the scaled
+     * price is 155000. Valid scale ranges are 0-10. It's not included if the last update transaction didn't include the
+     * BaseAsset/QuoteAsset pair.
+     */
+    Scale?: number
+  }
+}
